# 🚀 Modernize Portfolio: Migrate to Vite, Add Contact Form, and Optimize Performance