--- a/package.json
+++ b/package.json
@@ -4,13 +4,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emailjs/browser": "^4.4.1",
     "axios": "^1.7.9",
     "framer-motion": "^11.18.1",
     "gh-pages": "^5.0.0",
-
     "lucide-react": "^0.473.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hot-toast": "^2.5.2",
     "react-icons": "^4.4.0",
     "react-masonry-css": "^1.0.16",
     "react-router-dom": "^6.10.0",
@@ -25,7 +26,6 @@
     "build": "vite build",
     "preview": "vite preview"
   },
-
   "browserslist": {
     "production": [
       ">0.2%",
@@ -41,6 +41,9 @@
   "devDependencies": {
     "@vitejs/plugin-react": "^4.6.0",
     "autoprefixer": "^10.4.8",
+    "imagemin": "^9.0.1",
+    "imagemin-pngquant": "^10.0.0",
+    "imagemin-webp": "^8.0.0",
     "postcss": "^8.4.16",
     "tailwindcss": "^3.1.8",
     "vite": "^6.3.5"