--- a/src/components/OptimizedImage.jsx
+++ b/src/components/OptimizedImage.jsx
@@ -1,0 +1,66 @@
+import React, { useState } from 'react';
+
+const OptimizedImage = ({ 
+  src, 
+  webpSrc, 
+  alt, 
+  className = '', 
+  style = {},
+  placeholder = true,
+  ...props 
+}) => {
+  const [isLoaded, setIsLoaded] = useState(false);
+  const [hasError, setHasError] = useState(false);
+
+  const handleLoad = () => {
+    setIsLoaded(true);
+  };
+
+  const handleError = () => {
+    setHasError(true);
+    setIsLoaded(true);
+  };
+
+  return (
+    <div className={`relative overflow-hidden ${className}`} style={style}>
+      {/* Placeholder/Loading state */}
+      {placeholder && !isLoaded && (
+        <div className="absolute inset-0 bg-gradient-to-br from-slate-800 to-slate-700 animate-pulse flex items-center justify-center">
+          <div className="w-16 h-16 border-4 border-blue-500/30 border-t-blue-500 rounded-full animate-spin"></div>
+        </div>
+      )}
+      
+      {/* Optimized image with WebP support */}
+      <picture>
+        {webpSrc && <source srcSet={webpSrc} type="image/webp" />}
+        <img
+          src={src}
+          alt={alt}
+          onLoad={handleLoad}
+          onError={handleError}
+          className={`transition-opacity duration-500 ${
+            isLoaded ? 'opacity-100' : 'opacity-0'
+          } ${hasError ? 'filter grayscale' : ''}`}
+          loading="lazy"
+          {...props}
+        />
+      </picture>
+      
+      {/* Error state */}
+      {hasError && (
+        <div className="absolute inset-0 bg-slate-800 flex items-center justify-center">
+          <div className="text-center text-gray-400">
+            <div className="w-16 h-16 mx-auto mb-2 bg-slate-700 rounded-full flex items-center justify-center">
+              <svg className="w-8 h-8" fill="currentColor" viewBox="0 0 20 20">
+                <path fillRule="evenodd" d="M4 3a2 2 0 00-2 2v10a2 2 0 002 2h12a2 2 0 002-2V5a2 2 0 00-2-2H4zm12 12H4l4-8 3 6 2-4 3 6z" clipRule="evenodd" />
+              </svg>
+            </div>
+            <p className="text-sm">Image unavailable</p>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default OptimizedImage;