… src/assets/optimized/Portfolio_Image.jpg
… src/assets/optimized/Portfolio_Image.webp